# Faster runs via RuboCop::Runner