Every time we call `RuboCop::CLI#run` (which happens once per file being linted), it sets up a new RuboCop::Runner.
This is pretty slow, especially from `RuboCop::Options#parse` and `RuboCop::Runner#mobilized_cop_classes`.

We can skip all this repeated setup by using RuboCop::Runner directly. This also means we don't have to capture the stdout/stderr streams from CLI, since the relevant information is available directly on the runner instance.

Before:

```
626 files inspected, 0 lints detected

________________________________________________________
Executed in   23.22 secs    fish           external
   usr time   20.12 secs    0.32 millis   20.12 secs
   sys time    2.57 secs    2.49 millis    2.56 secs
```

After:

```
626 files inspected, 0 lints detected

________________________________________________________
Executed in   17.64 secs    fish           external
   usr time   15.30 secs    0.34 millis   15.30 secs
   sys time    1.93 secs    2.60 millis    1.93 secs
```

I've introduced a subclass of RuboCop::Runner to handle some of the customizations we need. ruby-lsp have a similar approach here - https://github.com/Shopify/ruby-lsp/blob/13bb8b0addf96248a951dca988e40127c86b6ba8/lib/ruby_lsp/requests/support/rubocop_runner.rb#L54, which might be useful for comparison.