@jdelStrother 2025-06-22
I've also got a [POC here](https://github.com/jdelStrother/haml-lint/commit/rubocop-cache) that makes use of RuboCop's cache, which dramatically improves times on already-processed files:

```
626 files inspected, 0 lints detected

________________________________________________________
Executed in    5.22 secs    fish           external
   usr time    4.30 secs    0.32 millis    4.30 secs
   sys time    0.88 secs    4.40 millis    0.87 secs
```

RuboCop doesn't use its cache on code from stdin, so when we generate ruby code from the haml file for RuboCop to lint, we could write it to, eg, `tmp/haml-lint/app/views/posts/index.html.haml.rb`.
Relying on a tmp directory doesn't seem great though, so maybe we can defer that idea to a future PR.